Move TriSolve to tri.h
