+++ src/slap/tri.c
@@ -146,3 +146,32 @@ enum slap_ErrorCode slap_MakeLowerTri(Matrix A) {
   }
   return SLAP_NO_ERROR;
 }
+
+enum slap_ErrorCode slap_TriSolve(Matrix L, Matrix b) {
+  SLAP_ASSERT_VALID(L, SLAP_INVALID_MATRIX, "LowerTriBackSub: L matrix invalid");
+  SLAP_ASSERT_VALID(b, SLAP_INVALID_MATRIX, "LowerTriBackSub: b matrix invalid");
+  SLAP_ASSERT(slap_NumCols(L) == slap_NumRows(b), SLAP_INCOMPATIBLE_MATRIX_DIMENSIONS,
+              SLAP_INCOMPATIBLE_MATRIX_DIMENSIONS,
+              "LowerTriBackSub: L has %d columns but b has %d rows", slap_NumCols(L),
+              slap_NumRows(b));
+  int n = b.rows;
+  int m = b.cols;
+  bool tL = L.mattype == slap_TRIANGULAR_UPPER || slap_IsTransposed(L);
+
+  for (int j_ = 0; j_ < n; ++j_) {
+    int j = tL ? n - j_ - 1 : j_;
+    for (int k = 0; k < m; ++k) {
+      sfloat* xjk = slap_GetElement(b, j, k);
+      sfloat Ljj = *slap_GetElement(L, j, j);
+      *xjk /= Ljj;
+
+      for (int i_ = j_ + 1; i_ < n; ++i_) {
+        int i = tL ? i_ - (j_ + 1) : i_;
+        sfloat* xik = slap_GetElement(b, i, k);
+        sfloat Lij = *slap_GetElement(L, i, j);
+        *xik -= Lij * (*xjk);
+      }
+    }
+  }
+  return SLAP_NO_ERROR;
+}
